backend/requirements.txt
# Web Framework
fastapi==0.115.11
uvicorn==0.27.0
gunicorn==21.2.0  # Production WSGI server
starlette==0.46.0
pydantic==2.10.6
pydantic_core==2.27.2

# Database
SQLAlchemy==2.0.38
alembic==1.14.1  # For database migrations
psycopg2-binary==2.9.9  # PostgreSQL driver (for production)

# Authentication
passlib==1.7.4
python-jose==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20  # For form data parsing

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Web Server Dependencies
anyio==4.8.0
h11==0.14.0
click==8.1.8
typing_extensions==4.12.2

# Networking & HTTP
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
sniffio==1.3.1
urllib3==2.3.0

# Supporting Libraries
annotated-types==0.7.0
async-timeout==5.0.1
blinker==1.9.0
cffi==1.17.1
cryptography==44.0.2
ecdsa==0.19.0
exceptiongroup==1.2.2
future==0.18.2
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
pyasn1==0.4.8
pycparser==2.22
six==1.15.0
zipp==3.21.0

# Date and Time Handling
python-dateutil==2.9.0.post0
pytz==2025.1
tzdata==2025.1

# Flask (optional, can be removed if not used)
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
Werkzeug==3.1.3

# Data Science/ML Libraries (can be removed if not used in production)
contourpy==1.3.0
cycler==0.12.1
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
geographiclib==2.0
geopy==2.4.1
joblib==1.4.2
kiwisolver==1.4.7
macholib==1.15.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
redis==5.2.1
regex==2024.11.6
rsa==4.9
spotipy==2.25.0
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
